trasferring all in a database for faster and simultaneous query. also adding a waiter_bot to send that bot in pause for checking new files

+++ scheduler.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
-time=86400 #24h
-
 #source .venv/bin/activate
+h=true
+t=true
 while true; do
-	python3.10 autocheck.py
-	python3.10 mainbot.py &
-	PID=$!
-	sleep $time
-	kill $PID
+	if $h; then
+		kill $PID
+		python3.10 mainbot.py &
+		PID=$!
+		h=false #se e' gia partito non deve ripartire
+	fi
+	if [ "$(date '+%H:%M')" = "01:00" ]; then	
+		kill $PID
+		python3.10 bot_pause_messages.py &
+		PID=$!
+		python3.10 autocheck.py
+		h=true #faccio ripartire il bot
+	fi
+	sleep 58 #<1 min in modo da beccare sempre
 done
